Add a 'label all' button to label all results in view
